+++ ElvUI_EltreumUI/Core/Options.lua
@@ -4156,36 +4156,27 @@ The Item Level shown on the Character Panel Skin uses code from Simple Item leve
 	ElvUI_EltreumUI.Options.args.changelog.args.added = ACH:Group(E:TextGradient("Added", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 3)
 	ElvUI_EltreumUI.Options.args.changelog.args.added.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.added.args.description = ACH:Description([[
-Added an option to hide the icon on the Teleports Datatext
-Added timer text to the Mirror Timers in Retail
-Added Delve Tier list to the Instance indicator
+Added an experimental instant DPS tag [eltruism:unitdps]
+Added more models to the Portrait Fix
 ]], 3, "small", nil, nil, nil, nil, "full")
 	ElvUI_EltreumUI.Options.args.changelog.args.updated = ACH:Group(E:TextGradient("Updated", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 3)
 	ElvUI_EltreumUI.Options.args.changelog.args.updated.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.updated.args.description = ACH:Description([[
-Updated Waypoint to attempt to solve the taints related to setting user waypoints
-Updated the Automatic Border adjust to account for Classic Era Totem Shamans
-Updated Unitframe models and also added Monk to Mists of Pandaria Classic
-Updated gradient Custom Glow to avoid using the wrong color
-Updated profiles to add a custom Power Prediction color
-Updated the Quest Item Bar to attempt to avoid taints
-Updated NPC Ids for Nameplate Classification Icons
-Updated list of models for the Portrait Fix
-Updated LFG skin to add the Leaver Icon
-Updated WindTools profile
-Updated Auctionator skin
-Updated Immersion skin
-Updated CVar list
+Updated some tags and other functions to allow AI party members to use class colors
+Updated Sockets in Classic Era to better show tooltips for the Enchants
+Updated Quest Item Bar to allow item usage in combat
+Updated Bag Profession Icons to add missing IDs
+Updated Immersion skin to better handle fonts
+Updated music for Earthen and Dracthyr
 ]], 5, "small", nil, nil, nil, nil, "full")
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed = ACH:Group(E:TextGradient("Fixed", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 4)
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed.args.description = ACH:Description([[
-Fixed an issue where the Inspect frame in Mists of Pandaria Classic was not being skinned correctly
-Fixed an issue where coloring the Waypoint wouldn't work if the Waypoint ETA was disabled
-Fixed an issue with the profile due to a change in the ElvUI tooltip database
-Fixed an issue where some AFK music was not being played in Classic Era
-Fixed an issue with the Vellum button on Classic
-Fixed an issue with ReforgedArmory
+Fixed an issue when disabling ElvUI's spellbook skin would still have the shadows
+Fixed an issue when using Leatrix Maps and having ElvUI world map disabled
+Fixed shadows on unitframes when the infopanel was enabled but not shown
+Fixed an error with the Item Delete due to a change on Blizzard's API
+Fixed an error in Classic Era when inspecting units
 ]], 7, "small", nil, nil, nil, nil, "full")
 
 	--[[